backend/src/main/resources/remote_db/requirements.txt
chromadb==0.5.5
dashscope==1.20.3
pyzmq==26.0.3